feat: Support receiving `2Dcur`s from multiple sources
Previously it was assumed that the `2Dcur`s would come from the same
single source. So the `source` element / message was skipped and if a
different source started sending messages, it would have to catch up in
frame number and if it did, it would fight with the original source to
be considered valid frames.