feat: Run relative to element not w/ abs coordinates
Also report clientX & clientY correctly by adjusting based on elements
client bounding rect. This allows touch events to work better in
libraries the rely on client coordinates to get coordinates relative to
the viewport like document.elementFromPoint(). This is actually correct
as clientX is viewport relative.

To support arbitrary regions `dimensionsToFakeElement()` was added to
create a fake element like object that fits the width, height and offset
of the original arguments for TuioToTouch's constructor.