feat: Account for scale transform
Does not support all transforms though may want to in the future. For
now scale is easily implemented by using the client bounding rectangle's
width and height which represent the resulting width and height of the
element after the transform.

Added a scroll to the document origin to `clearBody()` to reset after
scrolling in tests.