fix: Ignore non-bundled messages w/o source yet set
Not 100% sure this is how this should be handled, but there is nothing
to do with the messages if there is no source. Its assumed that messages
received before a `source` type message are caused by the `source`
message being dropped likely during the TUIO server boot up process. Aka
its a timing issue where a message is sent before the server is ready.